requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3

tensorflow==2.15.0
keras==2.15.0

numpy>=1.23.5,<2.0
Werkzeug>=2.3.8,<3.0
psycopg2-binary>=2.9
Pillow>=9.0.0
python-dotenv>=1.0
cloudinary>=1.32.0
gdown>=4.6.4
gunicorn==20.1.0
huggingface_hub==1.8.0